Probability density function of multiple structured Normal inverse Wishart

Probability density function of structured Normal inverse Wishart (sNiW) for multiple inputs, on the log scale.

Usage

mmsNiWlogpdf(U_xi, U_psi, U_Sigma, U_xi0, U_psi0, U_B0, U_Sigma0, U_df0)

Arguments

U_xi

a list of length n of observed mean vectors, each of dimension p

U_psi

a list of length n of observed skew vectors of dimension p

U_Sigma

a list of length n of observed covariance matrices, each of dimension p x p

U_xi0

a list of length K of mean vector parameters for sNiW, each of dimension p

U_psi0

a list of length K of mean vector parameters for sNiW, each of dimension p

U_B0

a list of length K of structuring matrix parameters for sNiW, each of dimension 2 x 2

U_Sigma0

a list of length K of covariance matrix parameters for sNiW, each of dimension p x p

U_df0

a list of length K of degrees of freedom parameters for sNiW, each of dimension p x p

Examples

hyperG0 <- list()
hyperG0$b_xi <- c(-1.6983129, -0.4819131)
hyperG0$b_psi <- c(-0.0641866, -0.7606068)
hyperG0$kappa <- 0.001
hyperG0$D_xi <- 16.951313
hyperG0$D_psi <- 1.255192
hyperG0$nu <- 27.67656
hyperG0$lambda <- matrix(c(2.3397761, -0.3975259,-0.3975259, 1.9601773), ncol=2)

xi_list <- list()
psi_list <- list()
S_list <- list()
for(k in 1:1000){
 NNiW <- rNNiW(hyperG0, diagVar=FALSE)
 xi_list[[k]] <- NNiW[["xi"]]
 psi_list[[k]] <- NNiW[["psi"]]
 S_list[[k]] <- NNiW[["S"]]
}
mmsNiWlogpdf(U_xi=xi_list, U_psi=psi_list, U_Sigma=S_list,
            U_xi0=list(hyperG0$b_xi), U_psi0=list(hyperG0$b_psi) ,
            U_B0=list(diag(c(hyperG0$D_xi, hyperG0$D_psi))) ,
            U_Sigma0=list(hyperG0$lambda), U_df0=list(hyperG0$nu))
#>    [1]  5.689243280  6.537536114  2.714334471  5.034696016  9.545230504
#>    [6]  8.672423475  8.867101426  8.545295598  6.426193216  7.287473763
#>   [11]  2.077063419  1.561722587  5.481030064  5.260418479  5.684044760
#>   [16]  1.460374983  7.034427789  4.531350472  7.726773164  5.161281596
#>   [21]  2.130520102  3.572350076  3.302560490  5.879554359  3.929409719
#>   [26]  7.170051757  7.127320746  5.908747521  5.785881828  5.855175117
#>   [31]  5.214468182  7.238991407  6.883319261 -0.202295421  9.054414731
#>   [36]  6.964188176  0.893865969  5.366851260  5.101558314  7.149961823
#>   [41]  6.212421055  4.980959096  3.372033248  6.731682469  1.175910036
#>   [46]  7.085495847  1.428996841  8.977171579  6.166102279  4.419235764
#>   [51]  6.446948216  5.172116387  6.952369348  7.716214016  3.008014432
#>   [56]  2.223314707  7.596897093  3.819820749  3.243957408  3.387562873
#>   [61] -2.228709816  7.816328289  4.960886341  6.300793692  6.820644847
#>   [66]  5.239343880  7.549238522  7.790165809  6.765079032  2.604068542
#>   [71]  8.955957420  3.619437672  7.039035144  7.578706443 -1.318171345
#>   [76]  3.286538448  4.656748023  6.345170744  8.877172112  1.449831924
#>   [81]  7.409239523  1.482496148  7.337245303  6.182724504  7.994471093
#>   [86]  2.505672777  6.107862629  5.634700651  7.786272250  8.124156496
#>   [91]  7.185841029  3.187968997  6.840443972  3.256552577 -0.939176223
#>   [96]  5.675863232  1.523874977  2.886908977  3.495673486  5.653078318
#>  [101]  5.893758533  6.936419004  7.882110653  7.591673154  6.297923741
#>  [106]  7.577236392  7.821623698  6.505721299  8.075756849  2.354342979
#>  [111]  3.262835332  7.660159495  9.557983322  3.582578213  2.396413362
#>  [116]  6.684284323  8.364391650  7.243891327  7.139147714  7.948695477
#>  [121]  8.167878262  7.496132994  7.144079477  8.789990148  7.293900460
#>  [126]  3.442916348  5.215325922  6.323679218  7.136288595  6.849324618
#>  [131]  6.065980870  4.062323097  5.674458409  1.099323827  4.863389470
#>  [136]  7.381567218  3.507132356  3.934384090  7.624698293  7.516606155
#>  [141]  8.280254644  8.091795192  5.160088669  7.966905543  7.371563791
#>  [146]  7.422356429  5.286717218  8.405937875  6.632071004  2.917921506
#>  [151]  7.655723650  6.173256171  6.197513415  7.413195969  3.737273420
#>  [156]  6.544916450  3.691601486  5.551304095  3.761636428  9.211102341
#>  [161]  6.087706226  2.486401945  7.548173394  7.291353399  6.138576779
#>  [166] -0.221697429  7.619834754  5.027953429  3.990409032  3.527740062
#>  [171]  7.027998490  4.836467126  6.737619566  5.910178797  3.391346319
#>  [176]  8.509173587  8.107038542  8.138672420  7.203110521  1.943937338
#>  [181]  8.795679900  8.553560301  6.463503142  7.661739578  7.366268106
#>  [186]  3.256349507  7.276287236  7.291667721  6.427142250  9.189153851
#>  [191]  7.043982717  4.316314521  6.839931314  8.193663521  6.228144634
#>  [196]  1.122449679  3.686621839  5.637492094  1.772780215  5.330102645
#>  [201]  3.164338072  5.197122608  6.569952955  4.190457232  7.805997200
#>  [206]  5.845252326  4.382315796  8.382205609  6.519024293  5.679950676
#>  [211]  5.843272600  6.753588737  8.349338614  6.670975491  6.830060275
#>  [216]  5.110551159  0.894803963  8.020712992  3.317188348  3.536240203
#>  [221]  3.731268300  7.307166737  2.313337760  4.708377171  5.662379368
#>  [226]  4.703746407  8.423259521  5.021399916  0.702716654  5.370214039
#>  [231] -2.139716296  6.118042919  6.800748280  8.591295209  2.659031173
#>  [236]  6.565173930  7.738564980  8.432149013  3.617543061  6.230087138
#>  [241]  6.643343019  4.912807994  8.355607104  4.544409384  6.757434370
#>  [246]  7.813415597  7.270458578  5.509474501  6.554453258  2.217252732
#>  [251]  5.207413530  5.629002096  2.863723369  6.756745337  2.725054679
#>  [256]  5.683185421  7.590322534  7.737589908  7.587360501  5.327693719
#>  [261]  6.790295433  4.148019922  7.856234954  1.078879224  8.921806835
#>  [266]  6.332786705  6.498915154  5.667329070  7.238571517  4.939245212
#>  [271]  5.848916277  5.692650104  7.386100366  3.016563992  7.224697873
#>  [276]  6.211686061  6.573554941  7.337405464  5.370968446  7.249758966
#>  [281]  5.840417731  8.543052701  8.366817101  8.398659265  7.385621433
#>  [286]  8.100156586  7.655436982  3.092533278  5.854772341  8.174012089
#>  [291]  8.849025652  4.197344032  7.854932693  4.044092656  4.907919256
#>  [296]  7.600330991  4.526311984  0.254606989  2.489919408  3.877147388
#>  [301]  5.280810028  8.028409634  6.500457258  6.686391158  6.753388654
#>  [306]  6.308719816  9.554337388  5.702108722  9.039334338  7.103097103
#>  [311]  8.566911483 -0.525488831  8.760537087  7.871736428  2.833330071
#>  [316]  6.445477009  9.174788214  6.498066777  7.463600630  7.769685750
#>  [321]  0.165797165  6.412940447  4.457652097  1.153889876  6.825231177
#>  [326]  3.141823193  7.946565466  4.526503931  4.656322249  6.957952759
#>  [331]  1.439347570  5.254980687  6.772682897  5.374747722  7.058085812
#>  [336]  4.616354388  6.269616057  7.334524024  3.920750345  7.312034260
#>  [341]  5.387172612  6.777076077  8.916543139  8.421333289  5.711057788
#>  [346]  7.376638075  6.942248273  4.509641288  7.429615931  8.924939303
#>  [351]  3.172611887  5.720875352  5.661573431  7.796671421  5.504805358
#>  [356]  6.762641383  2.918179066  7.491256913  4.856566003  6.567825807
#>  [361]  1.263517283  2.986281294  6.356226707  4.223916181  5.124053897
#>  [366]  7.416081253  7.747399374  4.047724557  4.777735802  6.615414724
#>  [371]  4.439296550  3.766112228  4.811589604  8.975459375  8.454758056
#>  [376]  1.808495301  1.625308847  4.597488552  8.968167170  6.337860695
#>  [381]  6.626589213  7.302943925  6.559766111  8.355344890  3.426207515
#>  [386]  5.770642032  0.367566465  4.283104695  0.340234233  6.541332742
#>  [391]  7.097097181  7.609325379  7.502126462  4.182359028  7.163204481
#>  [396]  4.159226502  5.318664414  6.411585368  8.249683375  3.347274475
#>  [401]  8.401826084  8.753792197  3.073620169  3.252322290  5.702591487
#>  [406]  6.355489799  6.615650505  7.505194200  1.444098655  3.126735805
#>  [411]  6.921388182  7.843396945  6.805538632  5.915302351  3.166617688
#>  [416]  5.397399301  7.221579515  7.689827745  0.290746907  6.544188833
#>  [421]  7.069537131  5.827078857  6.497975104  6.480786397 -0.152490235
#>  [426]  8.334267600  8.576646471  7.785966289  8.726738258  8.297315606
#>  [431]  7.760817339  4.283588356  4.413555275  7.530303119  6.164887132
#>  [436]  5.928350747 -0.904990782  2.706025671  5.617969260  3.853563809
#>  [441]  3.486676215  8.363717060  6.724902309  4.790311908  4.504640307
#>  [446]  6.888876349  7.085186344  7.678613861  1.656568323  6.460735180
#>  [451]  6.473354731  2.866765500  5.116231206  6.638637189  7.822449909
#>  [456]  1.826305345  4.952583964  6.275071192  8.227717458  6.416360316
#>  [461]  5.269759707  5.364834395  8.713343093  6.811341953  2.199081449
#>  [466]  7.810525430 -0.009389914  7.225538264  3.729472354  6.006223331
#>  [471]  3.122615879  7.275909161  5.436091626  6.492824247  3.923832057
#>  [476]  5.839573417  3.381002893  6.316209768  9.262214811  7.049953577
#>  [481]  5.085727332  7.726361988  4.693540932  9.300627879  5.550219938
#>  [486]  3.405555417  9.291213662  4.818048206  3.789720157  6.223069246
#>  [491]  9.201759510  8.280455988  4.009637084  1.194725318  7.938349008
#>  [496]  7.561978344  3.433920685  2.173174556  6.133695068  7.301467049
#>  [501]  7.090016076  8.045700550  4.892394419  7.162669032  2.364426049
#>  [506]  5.887971251  3.907544811  6.541506076  7.235655845  8.632867132
#>  [511]  7.777197837  6.578937608  7.837656309  5.465287701  2.954927696
#>  [516]  3.007010894  8.602758803  2.018923600  4.966433143  7.764380397
#>  [521]  8.570071359  3.315064181  2.144667892  4.091162663  5.031027476
#>  [526] -0.842998380  7.117157398  5.832338467  6.582508858  4.064905884
#>  [531]  8.822909060  5.542772513  6.670500181  6.568421080  6.895059232
#>  [536]  8.100553467  2.324390485  3.986745675  8.700207746  7.662654104
#>  [541]  7.033065891  4.700837646  8.193036144  4.235572935  4.669748784
#>  [546]  5.812173551  6.663092120  8.724990855  2.503613417  5.388924595
#>  [551]  1.532710769  8.745389739  5.589845191  3.341119281  5.148491008
#>  [556]  2.405634513  7.346800583  6.542627769  5.954268059  2.572505887
#>  [561]  2.945381348  3.885954910  8.054548826  6.670166119  6.463405765
#>  [566]  0.287251774  4.887878333  3.527537991  8.581924111  5.323341158
#>  [571]  6.605768785  6.960762954  6.280036116  1.184873639  2.195522929
#>  [576]  7.057933590  5.030246546  3.932741556  8.944845927  4.261583578
#>  [581]  2.170595318  3.768420660  8.055821346  7.058012295  7.303270729
#>  [586]  5.878905192 -0.200046211  6.565642627  7.758809591  7.081770278
#>  [591]  5.239830710  5.457920434  5.093236939  8.046077313  6.586291527
#>  [596]  6.589078665  7.069085929  6.491978470  7.637200356  6.964831996
#>  [601]  7.296895459  6.690531333  8.822452695  5.386904766  4.667885093
#>  [606]  8.443739271  5.069943296  5.395587907  7.522430795  5.220178257
#>  [611]  5.740419230  4.672943031  6.296707598  8.593212881  8.409981782
#>  [616]  1.734441632  5.620826549  7.318277772  3.300994106  9.390092091
#>  [621] -3.224947155  7.195333783  7.110367551  4.419259428  7.606218444
#>  [626]  9.137781143  7.618162637  5.403441477  4.762640043  5.206834822
#>  [631]  6.245878694  6.601269581  6.801241454  5.879705409  4.400671946
#>  [636]  6.968676200  7.812002471  8.398311551  3.331445813  7.160306862
#>  [641]  7.980979964  3.031171001  7.752411012  5.403135941  6.549017550
#>  [646]  4.019644001  3.942268858  2.183480625  7.881390800  6.872391677
#>  [651]  8.217726790  7.239194522  6.764632210  5.020807445  7.458553083
#>  [656]  0.131095330  3.983935170  7.688877293  8.949594421  3.560413531
#>  [661]  7.609080012  5.770313093  6.164637729  5.738192779  5.142738526
#>  [666]  6.026847914  8.207324346  2.806305915  4.522292903  8.212860765
#>  [671]  7.230538934  5.298506995  4.534339041  7.081479361  6.362947106
#>  [676]  5.264682012  7.230287728  3.347725729  5.631065854  7.353437371
#>  [681]  7.639633958  6.288040173  2.430412685  5.951077650  7.397847184
#>  [686]  7.807779079  6.175181257  5.603895753  5.412672701  2.605620969
#>  [691]  6.187306832  6.139340519  5.314597587  6.468041684  6.814555016
#>  [696]  8.476768364  2.471059619  2.863700953  6.556538185  4.606556562
#>  [701]  8.215602801  5.658607335  4.749897235  8.032265359  6.189713672
#>  [706]  6.698308565  8.132883792  6.969941725  3.339393520  5.262544529
#>  [711]  4.688127116  6.173615058  8.400437655  3.835696776  5.495948225
#>  [716]  8.567242263  2.226524947  6.376797344  5.179578989  3.966576196
#>  [721]  4.934836254  7.289454602  7.616586595  3.761281319  6.997625276
#>  [726]  6.880067302  7.269448907  7.715717261  6.248487947  6.621944135
#>  [731]  5.657518683  5.074483868  7.027000925  2.055521751  5.947083977
#>  [736]  2.969367152  7.950058152  7.527979673  6.879307882  5.053697443
#>  [741]  4.503030851  5.253210425  0.341578861  4.729542594  4.643459841
#>  [746]  5.863015334  6.943917160  7.829252135  3.865150907  7.839597735
#>  [751]  3.737367384  1.902437161  7.609952343  7.005866428  3.686916364
#>  [756]  7.627124445 -0.972017034  7.793216706  3.344919742  8.419571344
#>  [761]  2.557571480  6.268428081  5.779310735  5.396186186  8.382400552
#>  [766]  5.978826884  9.416179573  7.099443941  5.395045533  5.292700784
#>  [771]  3.631828312  4.726774656  6.037552461  1.328785401  7.738382588
#>  [776]  1.803663755  4.915675063  4.244097944  6.298386072  5.890145877
#>  [781]  4.710512361  5.901452822  8.464465932  7.261332932  8.991776974
#>  [786]  7.621259515  1.783190173  6.923763205  7.798133827  4.664287579
#>  [791]  6.738425627  7.513594388  4.664911709  4.911998513  3.602064465
#>  [796]  8.841752022  8.662892989  5.936422236  6.829928989  4.147228440
#>  [801]  6.545764162  7.079347136  1.795092211  5.840618500  8.846125706
#>  [806]  6.688214370  5.518976807  7.846831496  7.319251609  6.715955764
#>  [811]  7.353397768  5.204994006  3.166457546  6.196351988  9.740193764
#>  [816]  7.079878664  4.539694681  7.601001160  7.819979943  7.575435823
#>  [821]  5.713462323  4.334113379  6.808027414  5.386091589  5.966394219
#>  [826]  6.590018511 -1.061036678  8.445350557  3.326821393  7.168574766
#>  [831] -0.420953943  6.136652308  8.135528928  5.445872572  3.204211027
#>  [836]  6.564066457  3.707001135  8.018003201  4.090803591  2.450436111
#>  [841]  0.124356945  8.197853949  3.841965822  8.990024792  8.359341333
#>  [846]  2.751729695  4.296897545  1.939889168  6.886409794  7.975770931
#>  [851]  7.061346849  6.796059848  4.155972852  5.932596568  6.880463762
#>  [856]  5.169356660  5.724478357  7.972296232  6.249440167  6.381612805
#>  [861]  3.761382320  3.300213474  7.203821166  7.031051940  4.717042868
#>  [866]  6.459426650  4.662230577  1.629117797  3.905843120  4.397055581
#>  [871] -4.178393150  6.563841050  6.487052515  5.461131140  6.422169906
#>  [876]  4.752114399  6.899667807  7.337628733  4.026215654  6.243098113
#>  [881]  4.508995425  6.589483249  5.774583413  6.074448643  4.235056375
#>  [886]  5.812689176  7.664294402  7.110329171  4.643317963  8.113416565
#>  [891]  7.080867026  5.861173056  1.604480969  7.719142864  6.596932296
#>  [896]  6.488295391  7.313131729  6.306480551  5.140074385  7.600005658
#>  [901]  5.405405154  4.797453562  0.601362848  5.625301089  3.149922959
#>  [906]  6.975310374  7.724045470  0.529508535  7.691883283  8.433785208
#>  [911]  5.939551551  5.687634800  7.114004473  1.782721650  5.634711668
#>  [916]  4.663935288  3.778150075  9.112858922  8.428542321  6.335321390
#>  [921]  4.213324150  7.350717607  4.418541118  6.813474761  8.627154988
#>  [926]  5.823820705  7.062835423  7.998192983  8.565247221  5.571113038
#>  [931]  5.637800593  9.771050147  3.187446313  4.721353951  6.626176450
#>  [936]  7.655561787  1.638110885  5.674338005  7.063351102  4.928752203
#>  [941]  8.137402453  7.406527220  6.228311822  4.732998825  4.757473458
#>  [946]  7.039897057  2.596888738  0.969921273  6.084516362  5.884689330
#>  [951]  7.485446783  5.516912528  4.778680062  8.231666338  7.360048380
#>  [956]  2.835863626  2.956819552  5.391018418  6.561743067  6.226253545
#>  [961]  8.381079365  3.157036067  6.796582089  6.537255467  6.843261666
#>  [966]  4.217904715  6.944115384  7.261031968  8.286322402  0.348250787
#>  [971]  3.110753906  5.091804687  2.491682472 -3.950366099  5.489996159
#>  [976] -0.307741563  8.932782261  3.158383126  6.248195042  7.644891765
#>  [981]  5.743412594  6.394563063  6.132950046  6.588500616  7.792311427
#>  [986]  8.638594735  2.780447309  8.980831237  5.053221836  5.416357497
#>  [991]  6.799293180  5.489089543  5.807785568  6.059768800  5.896254467
#>  [996]  6.333249677  4.789924805  6.663543968  8.425502116  2.977118279

hyperG0 <- list()
hyperG0$b_xi <- c(-1.6983129)
hyperG0$b_psi <- c(-0.0641866)
hyperG0$kappa <- 0.001
hyperG0$D_xi <- 16.951313
hyperG0$D_psi <- 1.255192
hyperG0$nu <- 27.67656
hyperG0$lambda <- matrix(c(2.3397761), ncol=1)
#'xi_list <- list()
psi_list <- list()
S_list <- list()
for(k in 1:1000){
 NNiW <- rNNiW(hyperG0, diagVar=FALSE)
 xi_list[[k]] <- NNiW[["xi"]]
 psi_list[[k]] <- NNiW[["psi"]]
 S_list[[k]] <- NNiW[["S"]]
}

mmsNiWlogpdf(U_xi=xi_list, U_psi=psi_list, U_Sigma=S_list,
            U_xi0=list(hyperG0$b_xi), U_psi0=list(hyperG0$b_psi) ,
            U_B0=list(diag(c(hyperG0$D_xi, hyperG0$D_psi))) ,
            U_Sigma0=list(hyperG0$lambda), U_df0=list(hyperG0$nu))
#>    [1]  2.871116910  1.164065302 -0.042139274  1.121318232  1.915537613
#>    [6]  2.925285942  1.221807810  0.810960656  2.733623832  2.597308858
#>   [11]  2.566513407  2.351449274  1.636939915  0.933281745 -0.873135685
#>   [16]  2.213076141  2.922419168  0.286284328 -1.596313193  2.023418883
#>   [21]  2.556967623  0.432063075  1.981614656  1.507252286  2.799086768
#>   [26]  1.566383124  2.390404588  0.142950757  1.056436295  1.956613505
#>   [31]  2.448387498  2.058481245  2.806365081  2.526298779  2.674311551
#>   [36]  2.516916165  0.664003642  1.981238261  1.141881353  1.664687693
#>   [41]  2.898883649  0.528600230  0.625526181  1.218335291 -0.895131441
#>   [46]  2.553390854  2.539283862  2.162240748  2.430770424  2.576230098
#>   [51]  1.107686091  1.949145209  2.218607929  1.283645147  1.211992018
#>   [56]  2.290403285 -1.215908399  2.222638367 -1.259165368 -1.681948932
#>   [61]  2.887931686 -0.014509031  1.666032541  0.817387660  2.473254448
#>   [66]  2.427785697  0.250093679  2.033748261  1.949802894  1.626906819
#>   [71]  1.359001759  0.957598562  2.620321747  2.438952240  0.862169685
#>   [76]  2.828681902  1.335888562  0.833140562  0.141761186  0.492625804
#>   [81]  2.890428442  0.613442841  0.449974815 -2.232793383  1.198601904
#>   [86]  2.005700286  2.626020439  2.812351126  2.079489058  2.422671567
#>   [91]  0.427555560 -4.924038473  2.781584362  0.975669234  2.772240161
#>   [96]  2.563561184 -0.404859302  1.982259969  2.904904362 -0.951750287
#>  [101]  0.998784656  1.489081700  1.444209547  1.357346828  2.351973908
#>  [106] -0.736155729 -0.258931364  2.985308345  0.744968513  1.983328943
#>  [111]  1.813497092 -2.336983977 -0.531319712  1.313681921  1.650710980
#>  [116]  2.691510872  2.785142940  2.794279775  2.022565317  1.861164490
#>  [121] -0.455889142  2.150758594  2.288528580  2.246495794  2.309363531
#>  [126]  2.263404372  0.880647380  2.403322918  0.659505285  2.247633593
#>  [131]  2.588423803 -0.467166787  1.691861932  2.818750791  2.505710317
#>  [136] -3.044044032  1.495265981  1.066891672  1.972929116  2.383529160
#>  [141]  0.342727172  2.049688674  2.577453909  2.321722928  2.544767234
#>  [146]  1.425708460  2.300862946  1.310355137  2.172394204  1.149239745
#>  [151]  2.892243951  2.307383556  2.317649369  0.748521088 -0.044366309
#>  [156]  2.406257341  2.003843244  1.970450912 -0.090641628  2.602278325
#>  [161]  1.618568633  1.403725805  2.335545531  1.978332623  2.504197208
#>  [166]  2.219942469  2.138778785  2.717951083  2.658833942  2.254181333
#>  [171]  1.505792904  1.380959934 -1.027911115  2.788140053  1.248202145
#>  [176]  1.059822855  2.577257398 -0.906218836  1.160891430  0.932238941
#>  [181]  0.931541854  2.654215219 -2.043068018  2.209819378  0.996586290
#>  [186]  2.569690887  1.797201930  0.327488534  2.326262259  0.431969346
#>  [191]  1.912177088  1.124934370  2.280208246  2.534667398  1.967408230
#>  [196]  3.010596834  2.980425845  2.641666633  0.090756295  2.282246399
#>  [201]  1.598810499  0.291530050  0.216687166  2.065579317  2.028522550
#>  [206]  2.531637282 -0.088689898  1.764148273  1.427557665  0.684232672
#>  [211]  1.916863733  2.799562104  1.874165985  1.348156869 -0.009556799
#>  [216] -0.468242157  2.014417901  1.953169282  2.077700914  2.746577773
#>  [221]  2.238718282  2.301456501 -0.299696539  0.364861755  2.943817873
#>  [226]  1.763129053 -2.260844626 -3.700777421  1.720405723  2.213535228
#>  [231]  1.974151510 -0.364983137  2.395400011 -0.714812862  0.748377231
#>  [236]  2.618945889  2.403991946  0.319329520  2.397639350  2.711176317
#>  [241]  2.761591575  2.613393900  2.629689191  1.164472977  2.321317273
#>  [246]  1.231710790  2.469369017  1.606300275 -0.067291310  0.061436989
#>  [251]  1.026718005  2.079862685  1.836286379  1.957146785 -0.835891784
#>  [256]  2.944414740  2.216478236  2.694236254 -0.535615607 -0.885773295
#>  [261]  2.031934133  0.783638703 -1.022856647  0.792312168  1.879541279
#>  [266]  2.260175213  1.336621757  2.451869936  2.168775010  1.071194027
#>  [271]  1.017796019  0.500942186  2.110975422  1.640916870 -1.255086578
#>  [276]  2.707538882  0.098942585  2.906793215  2.124397794  1.595706552
#>  [281] -0.780335951  1.373661812  1.025621098  2.097970242  0.919450114
#>  [286]  1.618255881  2.648336558  2.849714034  2.581080324 -0.770702851
#>  [291]  2.759568002  0.355562039  1.983239234 -0.116286302  1.580750077
#>  [296]  2.049017001  2.411261591 -0.239171474  2.630217474 -2.590913554
#>  [301]  2.165772612 -3.893217529  0.229612054  0.284118942  1.997726654
#>  [306]  1.195436153  1.631516251  2.806567829  2.542829896 -3.163129889
#>  [311] -1.622958899 -2.348831108  1.375629956  0.187795809  2.534822813
#>  [316]  2.543450549  1.440374509  1.975156889 -4.049980160  1.979934195
#>  [321]  2.300399105 -0.346609223  2.783264496  1.302190562 -2.658992440
#>  [326]  1.566325078  0.304456066  2.911218184  2.931791991  1.653001373
#>  [331]  2.102999787  1.318857445  0.799480723  2.235093170  1.416420974
#>  [336]  0.231064273  1.013942882  1.958711771  1.971163378  2.913448300
#>  [341]  1.452877496 -0.060783609  0.301911404  1.732259395  2.069994518
#>  [346] -0.715758599 -2.420712656  1.728219252  0.788350635  1.214448896
#>  [351]  2.689825936  1.521690667  2.114270210  2.633787353  2.239735070
#>  [356]  2.245495595  2.374244857  0.786998565  1.496150439  2.014203488
#>  [361]  2.242024606  0.995527815  0.390263573  1.221866673  2.824831028
#>  [366]  2.439652404 -4.187539178  2.547968764  0.811682225  2.950327752
#>  [371]  1.777440095  1.652882965  1.732990681  0.850156784  0.505860262
#>  [376] -1.423252407  2.546729830 -0.721045106  2.689563194  2.626108500
#>  [381]  2.327481048  1.970566314  1.320197460  1.617557104  1.331272677
#>  [386]  1.803204984  2.304816528  2.739036131  0.741028955  2.059534788
#>  [391]  2.753902060  2.565335516  2.113742319  2.654173635 -1.067808909
#>  [396]  2.402469463  1.413930941  2.909220111 -0.826158742 -0.616251369
#>  [401]  1.443197771  0.797830255  1.159069940  2.756181496  1.538738541
#>  [406]  2.663191831  0.738625224  0.858274907  0.375596258  2.671193035
#>  [411]  0.102053671 -2.027477363  2.511860847  2.012106898  1.512428923
#>  [416]  2.704405327  2.069909475  1.660805976  1.944617961  0.450665521
#>  [421]  2.154308104  1.935491108  1.897355752 -1.456168101  2.250162608
#>  [426]  2.841547072  0.039480611  0.999433206 -1.171616628 -1.745975594
#>  [431]  2.108539433  0.982816399  1.606445710  2.485012662  2.277415139
#>  [436]  1.321747456  0.896779677  1.366672329  0.241310374  3.006290116
#>  [441]  1.339362434  0.546758139  2.818651280  2.541089951  2.685602797
#>  [446]  0.853087338  2.429836354  2.494439744  1.981061456  2.432789634
#>  [451] -0.566769431  2.331753924  1.260527742  0.127028348  1.536416508
#>  [456]  1.881699889  1.665974150  1.136368162  2.505627749  2.551257478
#>  [461]  2.067471353 -0.385941015  0.108092602 -1.325815534  1.651815607
#>  [466]  2.153292091  2.045477091  0.304369762  1.870936464  2.573108277
#>  [471]  1.527267994  1.817860246 -1.176146986 -0.620993851 -1.641284229
#>  [476] -1.155409064 -1.598403234  2.713492668  2.628182398  0.767185436
#>  [481] -3.574695278  2.654016952  1.195148913  2.669534976  1.807291965
#>  [486]  2.519010196  2.189920490  1.305850329  1.983809595  2.189041564
#>  [491]  2.699413565  2.806262179  1.783635430  2.136409669 -1.835842736
#>  [496]  2.870433037  1.940838726  2.132418118  1.880406466  1.902934493
#>  [501]  0.997427296  2.658652183  1.181091836  2.932100558  1.517775375
#>  [506]  1.937994532  1.268451186  0.088057451  1.558451995  2.932131248
#>  [511]  1.905716106  1.593768069 -0.424284406  2.331190726  1.943110213
#>  [516]  0.875723976 -3.296975173  1.935827494  0.692256928  1.798668404
#>  [521] -3.976721436  1.978334865  2.250273834  1.851869354  0.216366035
#>  [526]  2.644353257  2.885651084  2.093406192  0.103386091 -1.184438082
#>  [531]  2.396905435  1.515158248  0.342545063  0.364098658  1.127579544
#>  [536]  2.053946302  1.783671555 -0.963219314  0.113811720 -0.568214589
#>  [541] -0.931277016 -0.422131483  1.325371261  2.985955784  2.051911950
#>  [546]  2.785304638  0.988363522  0.117343246 -1.051105581  1.586379205
#>  [551] -0.028623054 -0.246518558  2.451591311  1.212427478  0.616692880
#>  [556]  0.842636693  0.233755375  1.703193424  2.819246887  2.762978666
#>  [561]  2.896018571  2.636886020 -4.072731286  2.828082377  1.335286534
#>  [566]  1.195144273  2.008904793  1.534507546  3.015971556 -0.260194185
#>  [571]  2.426685876  2.589926376  1.267963831 -0.618321031  1.912901129
#>  [576] -0.726727911  2.982718939  0.829838458  1.778178856  1.993648640
#>  [581]  1.742542672  2.102314555  2.850280310  0.915430848  2.314287846
#>  [586]  1.814726190  2.446391389 -1.162911853  1.849525516  2.210040990
#>  [591]  1.582469418  1.396827139  1.473477655 -1.345155389  2.149283196
#>  [596]  0.660442847  2.680840533  0.937494898 -0.908404179  2.908089730
#>  [601]  2.218250730  1.742489058  0.061394006  1.788009092  1.282647175
#>  [606]  1.318638576  0.716793856  0.732304494  1.845319546  2.547208646
#>  [611]  2.513677491  1.521476463  1.913774825  2.399639513  2.708249139
#>  [616]  0.265300397  2.386439865 -0.505018518  1.710299787  1.743076941
#>  [621]  0.473939029  2.843623240  2.209322341  2.579813226  2.208744690
#>  [626]  1.619151942  2.284539509  1.646452119  0.763689250  2.585342536
#>  [631]  2.118141827  1.719622539  1.430170811  2.747705427  2.793723379
#>  [636]  1.669184066  1.653171615 -2.472657417  1.866424504  1.455538190
#>  [641]  1.058034854  1.846485943  0.431147108  0.999010073  2.447136879
#>  [646]  2.361502108 -2.697685300  1.476455497  1.782107693  1.489820985
#>  [651] -0.542696233  2.140402007  2.223417471  2.688619849  2.611072311
#>  [656] -0.285039170  1.384523196  2.438091544  2.770791297  2.606282998
#>  [661]  2.938128232  2.789174981  2.878276299  1.053305559  1.731104970
#>  [666]  0.264233168  1.908603911  1.801740404  2.784749364  2.566311764
#>  [671]  1.524563485  1.023669030  1.266274090  0.541599723  0.313684092
#>  [676] -0.286431528 -0.733750928 -1.740362291  2.693252821  0.556550403
#>  [681]  2.479919061  1.481674172  2.261176512  2.780571621  2.189672861
#>  [686]  2.116376387 -5.020360333  2.833647997  2.056840163  1.160724083
#>  [691]  1.669226601 -0.173962880  1.781808879 -1.904825505  0.299976981
#>  [696]  1.661313529  2.471685711  2.411047619  1.587904627  1.607512300
#>  [701]  1.525163047 -1.147196930  2.340241451  1.420482883  1.310967278
#>  [706]  2.262495792  2.280789341  1.167009882  1.829059219  1.032744848
#>  [711]  2.799940623  1.607086689  2.248758388  2.294969767 -0.351524991
#>  [716] -4.503359624  1.856251099  1.667162922  1.465520686  2.724032926
#>  [721]  2.792429435  2.836053828  1.791367072  2.836994465  1.147403660
#>  [726] -2.599373470  1.864594502  1.661909294  0.419500182  1.459303554
#>  [731]  2.609266646  2.473300964 -2.629194612 -0.470530225  2.870483772
#>  [736] -0.584527818  2.037258681  2.505141546  1.677700958  0.356373470
#>  [741]  0.851601840  2.273144252  1.942204835  2.392848071  0.077936233
#>  [746]  1.399909621  0.213388920  2.098098531  1.453139195  0.508432822
#>  [751]  0.657917491  2.178996755  1.341571733  2.082954864  2.457390926
#>  [756]  0.931840779  2.549547575  1.442852258  2.554478042  1.777037533
#>  [761] -0.513389935 -2.638645712  1.926680908  2.331000525  2.680610858
#>  [766]  1.516520214  2.361923692  2.987144547  1.410371655 -1.651671185
#>  [771]  1.953741102 -0.444672175  0.756466966  0.243527026  1.416925126
#>  [776]  2.760804810  2.541929371  0.546804523  1.987535104  1.796850202
#>  [781]  1.730550767  1.263984715  1.718777933  2.913150783  0.430235311
#>  [786] -2.594677130  1.512634342  2.516401034  1.664924810  1.818363144
#>  [791]  2.434121005  1.296887450 -0.607499824  2.325647225  1.204393772
#>  [796]  2.102347819  2.733846898  0.337568438  0.833849306  2.115166699
#>  [801]  1.963385682  1.840510719 -1.101916765  1.493099351  2.053216162
#>  [806]  1.492416312  1.825680812  2.038135102  2.877921593  2.423276235
#>  [811]  2.785987623  2.623200234 -1.410558257 -2.483445403  2.701663000
#>  [816]  1.962161460  2.185405353  1.507663210  1.048002417  1.956039388
#>  [821]  2.181570265  0.828321712  0.668192270  1.660397117  2.394752488
#>  [826]  0.088893485  0.245975815 -1.179627386  2.231532493  0.736509437
#>  [831]  1.814178265 -4.103989740  0.392456053  1.842630093  1.103548536
#>  [836]  1.719493997  1.409181440  2.283453572  1.016476009  0.338862939
#>  [841]  2.397160158  2.341232594  2.842266126  2.732113005 -0.367626092
#>  [846]  1.905455500  2.654534195  1.021420805  1.551194711  2.765761967
#>  [851] -0.061838792  2.069658550  0.989402256  1.948440750  1.888545624
#>  [856]  1.780201921  2.382474822  2.627472841  2.232907843  2.270721679
#>  [861]  1.743436715  2.855044302  2.821632861  2.468345780  2.024210336
#>  [866]  2.235075531  2.417341754  0.637830887  1.907621832 -0.451240503
#>  [871]  2.559979024 -0.719477862  2.717998104  2.541055691  2.428884938
#>  [876]  0.337657183  2.002175698  1.752197025  1.600542640 -1.533534306
#>  [881]  0.279162499  2.741517612 -0.958433819  1.336112585  1.690040524
#>  [886]  0.451787241  2.886987962  2.268741701  2.101829175 -0.582496847
#>  [891]  1.672258788  2.268757824  0.147126591  0.723264128  2.331535459
#>  [896]  2.289519671  0.031294849 -2.467061302  1.796464936  2.515780890
#>  [901]  1.577385942  2.335547924  2.922991682 -0.271043076  2.671833408
#>  [906]  1.815019279  0.590464578  1.290614854  1.095054594  1.185197451
#>  [911]  0.484433344  2.107456446  0.842602627  2.266094666 -0.254906265
#>  [916]  2.161855327 -1.115203308  2.299397685  2.146424463  2.816934261
#>  [921] -0.964136698  0.598154178  2.294275322  1.578979458  2.770307910
#>  [926]  1.258211367  3.001519185  2.775261048  2.113276708 -1.323198501
#>  [931]  0.422879000  2.697013646  0.830819776  2.545402760  2.487647862
#>  [936]  2.674038707  0.970101008  0.812646369 -1.216799106  2.252396583
#>  [941]  2.353309742  2.164283015  0.416853659 -0.049733281  0.280131000
#>  [946]  0.781359304  0.991440964  1.234283898  1.309497909 -0.568653719
#>  [951]  2.526208628  1.923223643 -0.525080334  1.890238407 -0.707507184
#>  [956]  2.410172304  2.029714024  2.498869488  2.970567330 -0.867130850
#>  [961]  0.347151228  1.346623267 -1.045380606  0.570445454  1.114472748
#>  [966]  1.879747853  2.375683628 -1.612098621  1.847696878  2.769574928
#>  [971]  0.252917137  1.912650574  2.462870381  2.935004076  2.185000514
#>  [976]  1.765057480  2.195086887  1.563311590  0.986990241  2.493392190
#>  [981] -0.341869731  2.523040429  1.946603755  2.245762526  0.742687542
#>  [986]  2.444120811  2.568770480  0.826523449  1.789031637  2.387431327
#>  [991]  2.238121768  1.731182796  1.816183367  2.626779145  2.611494048
#>  [996]  2.791274561  2.307493125  1.765595691  2.144054579  1.594237432